HasC-Compiler/requirements.txt
pyyaml
svgwrite        # optional: SVG-Erzeugung mit Komfort
networkx        # optional: Netze automatisch layouten
cairosvg        # optional: SVG zu PDF/Gerber konvertieren
